+++ NEWS.md
@@ -1,5 +1,7 @@
 # angstroms dev
 
+* Added 'create_transport' function to flesh out details of a hydrodynamic forcing file for Atlantis. 
+
 * add better romshcoords logic, and romsdepth alias (simple arg provides original, 
 probably wrong, implementation)